requirements.txt
langgraph
langchain-core
langchain-community
langchain_openai
redis
pylint
mypy
GitPython
python-dotenv
faiss-cpu
flask
flask-cors
requests